[trunk] Ampliada documentación de Doxygen del motor de combate

+++ trunk/src/iniciar-objetos.cpp
@@ -38,7 +38,7 @@
 
 using namespace std;
 
-Uint8 crearFicheros(){
+Uint32 crearFicheros(){
     AtributoBase atrDentacosJoe(90,18,14,9,12,10,10,11,9,10,0.9,0.6,0.8,0.7,0.7,0.8,0.6,0.65);
     AtributoBase atrBaldos(120,8,7,12,9,10,14,8,12,6,0.7,0.85,0.6,0.9,0.5,0.6,0.9,0.5);
     AtributoBase atrGraimito(50,40,10,8,12,9,8,14,7,14,0.8,0.55,0.7,0.5,0.6,0.9,0.6,0.9);
